pyproject.toml
[project]
name = "amrt"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "L.Lu", email = "luliangce@gmail.com" }
]
dependencies = [
    "typer>=0.12.3",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/amrt"]

[[tool.rye.sources]]
url = "https://mirrors.bfsu.edu.cn/pypi/web/simple"
name = "bsfu"
